move string utility methods to Restaurant
